+++ mariadb/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Minor tweaks to ignore minor errors, also testing master-slave and master-master works
 * Add --disable-log-bin in mysql_install_db_args in mysql-server rc script
 * Fix error in changing mysql-server rc script
+* Add GALERAHOST parameter to set an access rule for a loadbalancer host
 
 ---